kyuubi jdbc code refactor
