+++ kyuubi-hive-jdbc/src/test/java/org/apache/kyuubi/jdbc/hive/UtilsTest.java
@@ -18,7 +18,6 @@
 
 package org.apache.kyuubi.jdbc.hive;
 
-import static org.apache.kyuubi.jdbc.hive.Utils.JdbcConnectionParams;
 import static org.apache.kyuubi.jdbc.hive.Utils.extractURLComponents;
 import static org.junit.Assert.assertEquals;
